Reorganised test cases.
