+++ tests/test_controllers_misc.py
@@ -1,0 +1,88 @@
+from loguru import logger
+
+from tests.conftest import OUT_DIR
+from vimms.Common import DEFAULT_ISOLATION_WIDTH, POSITIVE, set_log_level_warning
+from vimms.Controller import FixedScansController
+from vimms.Environment import Environment
+from vimms.MassSpec import IndependentMassSpectrometer
+
+
+class TestMultipleMS2Windows:
+    def test_data_generation(self, two_fixed_chems):
+        assert len(two_fixed_chems) == 2
+        assert two_fixed_chems[0].mass == 100
+        assert two_fixed_chems[1].mass == 200
+        assert len(two_fixed_chems[0].children) > 0
+        assert len(two_fixed_chems[1].children) > 0
+
+    def test_acquisition(self, two_fixed_chems):
+        mz_to_target = [chem.mass + 1.0 for chem in two_fixed_chems]
+        schedule = []
+        # env = Environment()
+        isolation_width = DEFAULT_ISOLATION_WIDTH
+        mz_tol = 0.1
+        rt_tol = 15
+
+        min_rt = 110
+        max_rt = 112
+
+        ionisation_mode = POSITIVE
+
+        controller = FixedScansController()
+        mass_spec = IndependentMassSpectrometer(ionisation_mode, two_fixed_chems, None)
+        env = Environment(mass_spec, controller, min_rt, max_rt)
+
+        ms1_scan = env.get_default_scan_params()
+        ms2_scan_1 = env.get_dda_scan_param(mz_to_target[0], 0.0, None, isolation_width, mz_tol, rt_tol)
+        ms2_scan_2 = env.get_dda_scan_param(mz_to_target[1], 0.0, None, isolation_width, mz_tol, rt_tol)
+        ms2_scan_3 = env.get_dda_scan_param(mz_to_target, [0.0, 0.0], None, isolation_width, mz_tol, rt_tol)
+
+        schedule = [ms1_scan, ms2_scan_1, ms2_scan_2, ms2_scan_3]
+        controller.set_tasks(schedule)
+        set_log_level_warning()
+        env.run()
+        assert len(controller.scans[2]) == 3
+
+        n_peaks = []
+        for scan in controller.scans[2]:
+            n_peaks.append(scan.num_peaks)
+
+        assert n_peaks[0] > 0
+        assert n_peaks[1] > 0
+        assert n_peaks[2] == n_peaks[0] + n_peaks[1]
+        env.write_mzML(OUT_DIR, 'multi_windows.mzML')
+
+
+class TestFixedScansController:
+    """
+    Tests the FixedScansController that sends a scheduled number of scans
+    """
+
+    def test_FixedScansController(self, two_fixed_chems):
+        logger.info('Testing FixedScansController')
+        mz_to_target = [chem.mass + 1.0 for chem in two_fixed_chems]
+        isolation_width = DEFAULT_ISOLATION_WIDTH
+        mz_tol = 0.1
+        rt_tol = 15
+        min_rt = 110
+        max_rt = 112
+        ionisation_mode = POSITIVE
+
+        controller = FixedScansController(schedule=None)
+        mass_spec = IndependentMassSpectrometer(ionisation_mode, two_fixed_chems, None)
+        env = Environment(mass_spec, controller, min_rt, max_rt)
+
+        ms1_scan = env.get_default_scan_params()
+        ms2_scan_1 = env.get_dda_scan_param(mz_to_target[0], 0.0, None, isolation_width, mz_tol, rt_tol)
+        ms2_scan_2 = env.get_dda_scan_param(mz_to_target[0], 0.0, None, isolation_width, mz_tol, rt_tol)
+        ms2_scan_3 = env.get_dda_scan_param(mz_to_target[0], 0.0, None, isolation_width, mz_tol, rt_tol)
+        schedule = [ms1_scan, ms2_scan_1, ms2_scan_2, ms2_scan_3]
+        controller.set_tasks(schedule)
+        set_log_level_warning()
+        env.run()
+
+        assert len(controller.scans[1]) == 1
+        assert len(controller.scans[2]) == 3
+        for scan in controller.scans[2]:
+            assert scan.num_peaks > 0
+        env.write_mzML(OUT_DIR, 'fixedScansController.mzML')
